Move module mocking into the unit_tests.utils module
This is so that the mocking of importable modules can be shared between
the test_tox_hooks_3 and (soon to be) test_tox_hooks_4 test modules.
